// src/app/store/action-types/adx-search.action-types.ts
export enum AdxSearchActionTypes {
  TOGGLE_DATA_SOURCES_FILTER = '[ADXSearch] Toggle Data Sources Filter',
  ADD_TERMS = '[ADXSearch] Add Terms',
  REMOVE_TERMS = '[ADXSearch] Remove Terms',
  ADD_COLUMNS = '[ADXSearch] Add Columns',
  ADD_CATEGORY = '[ADXSearch] Add Category',
  ADD_DATE_LIMIT_START = '[ADXSearch] Add Date Limit Begin',
  ADD_DATE_LIMIT_END = '[ADXSearch] Add Date Limit End',
  ADD_RANGE_RESULT = '[ADXSearch] Add Date Range Result',
  ADD_COMMAND_QUERY = '[ADXSearch] Add Command Query',
  SET_IS_LOADING = '[ADXSearch] Set is Loading',
  LOAD_ADX_SEARCH_REQUEST= '[ADXSearch] Load ADXSearch Request',
  LOAD_ADX_SEARCH_SUCCESS = '[ADXSearch] Load Adx Search Success',
  LOAD_ADX_SEARCH_ERROR = '[ADXSearch] Load Adx Search Error',
  REMOVE_ADX_SEARCH = '[ADXSearch] Remove Adx Search Success',
  LOAD_TIMELINE_SUCCESS = '[ADXSearch] Timeline Log Success',
  LOAD_TIMELINE_ERROR = '[ADXSearch] Load Timeline Error',
  TOGGLE_TIMELINE = '[ADXSearch] Toggle Timeline',
  LOAD_HISTOGRAM_ERROR = '[ADXSearch] Load Histogram Error',
  LOAD_HISTOGRAM_SUCCESS = '[ADXSearch] Load Histogram Success',
  TOGGLE_HISTOGRAM = '[ADXSearch] Toggle Histogram',
  SET_MAXIMUM_RESULT = '[ADXSearch] Set Maximum Result',
  REMOVE_MAXIMUM_RESULT = '[ADXSearch] Remove Maximum Result',
  SET_CLIENT_REQUEST_ID = '[ADXSearch] Set Client Request ID',
  SET_UNSUBSCRIBE = '[ADXSearch] Set Subscritpion',
  ADD_HISTOGRAM_FILTER = '[ADXSearch] Add Histogram Filter',
  REMOVE_HISTOGRAM_FILTER = '[ADXSearch] Remove Histogram Filter',
  RESET_FILTER = '[ADXSearch] Reset Filter',
  SET_CURRENT_FEED = '[ADXSearch] Set Current Feed',
  SET_IS_LOADING_TAB = '[ADXSearch] Set Tab is Loading',
  REMOVE_IS_LOADING_TAB = '[ADXSearch] Remove Tab is Loading',
  ADX_SERCH_DESTROYED = '[ADXSearch] ADX Search Destroyed',
  SET_ADX_SEARCH_DATE_RANGE = '[ADXSearch] ADX Search Date Range Set',
}